@@ -18,7 +18,8 @@
     "lint": "eslint --fix .",
     "prettier:check": "prettier --check '**/*.{js}'",
     "validate": "run-s lint prettier:check",
-    "build": "rollup -c"
+    "build": "rollup -c",
+    "prepare": "husky install"
   },
   "repository": {
     "type": "git",
@@ -27,7 +28,8 @@
   "author": "privateOmega <privateOmega@protonmail.com>",
   "contributors": [
     "amrita-syn <amrita.yadav@synapsica.com>",
-    "charuthaB <charutha.bhaskar@synapsica.com>"
+    "charuthaB <charutha.bhaskar@synapsica.com>",
+    "hanagejet"
   ],
   "license": "MIT",
   "bugs": {
@@ -35,55 +37,40 @@
   },
   "homepage": "https://github.com/privateOmega/html-to-docx#readme",
   "devDependencies": {
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
-    "@rollup/plugin-commonjs": "^12.0.0",
+    "@commitlint/cli": "^13.1.0",
+    "@commitlint/config-conventional": "^13.1.0",
+    "@rollup/plugin-commonjs": "^20.0.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^8.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^2.5.1",
-    "husky": "^4.2.5",
-    "lint-staged": "^10.2.11",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.5",
-    "rollup": "^2.18.2",
+    "@rollup/plugin-node-resolve": "^13.0.4",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.0",
+    "lint-staged": "^11.1.2",
+    "prettier": "^2.4.1",
+    "rollup": "^2.56.3",
     "rollup-plugin-cleaner": "^1.0.0",
     "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-terser": "^6.1.0",
-    "standard-version": "^8.0.0"
+    "rollup-plugin-terser": "^7.0.2",
+    "standard-version": "^9.3.1"
   },
   "dependencies": {
     "color-name": "^1.1.4",
     "escape-html": "^1.0.3",
     "html-to-vdom": "^0.7.0",
-    "image-size": "^0.8.3",
-    "jszip": "^3.5.0",
-    "shortid": "^2.2.15",
+    "image-size": "^1.0.0",
+    "jszip": "^3.7.1",
+    "lodash": "^4.17.21",
+    "nanoid": "^3.1.25",
     "virtual-dom": "^2.1.1",
     "xmlbuilder2": "2.1.2"
-  },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-      "pre-commit": "lint-staged"
-    }
   },
   "lint-staged": {
     "src/**/*.js": [
       "prettier --write",
-      "eslint --fix",
-      "git add"
+      "eslint --fix"
     ]
   }
 }
